qa/tasks/vstart_runner: do not send SIGTERM if no matched pid
otherwise following error is expected in some cases:

INFO:__main__:Traceback (most recent call last):
INFO:__main__:  File "/home/jenkins-build/build/workspace/ceph-api/qa/tasks/mgr/test_dashboard.py", line 18, in setUp
INFO:__main__:    self._assign_ports("dashboard", "ssl_server_port")
INFO:__main__:  File "/home/jenkins-build/build/workspace/ceph-api/qa/tasks/mgr/mgr_test_case.py", line 197, in _assign_ports
INFO:__main__:    cls.mgr_cluster.mgr_stop(mgr_id)
INFO:__main__:  File "/home/jenkins-build/build/workspace/ceph-api/qa/tasks/mgr/mgr_test_case.py", line 30, in mgr_stop
INFO:__main__:    self.mgr_daemons[mgr_id].stop()
INFO:__main__:  File "../qa/tasks/vstart_runner.py", line 558, in stop
INFO:__main__:    os.kill(pid, signal.SIGTERM)
INFO:__main__:TypeError: an integer is required (got type NoneType)

+++ qa/tasks/vstart_runner.py
@@ -552,6 +552,8 @@ class LocalDaemon(object):
             return
 
         pid = self._get_pid()
+        if pid is None:
+            return
         log.debug("Killing PID {0} for {1}.{2}".format(pid, self.daemon_type, self.daemon_id))
         os.kill(pid, signal.SIGTERM)
